requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
boto3==1.26.32
botocore==1.29.32
Deprecated==1.2.13
Django==4.1.2
django-address==0.2.8
django-cors-headers==3.13.0
django-phonenumber-field==7.0.1
django-redis==5.2.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.1.0
gunicorn==20.1.0
jmespath==1.0.1
packaging==21.3
Pillow==9.2.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
redis==4.3.4
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
wrapt==1.14.1
